Refactor webserver views
